Accumulator for messages
Messages had been duplicated as there are onMessage events for cached
as well as newly persisted messages. The accumulator is going to care
for that.

Also name had been in a cycle of changes as new instances of the object
literal were set in a cycle. It's gotten its own useEffect Hook now.