There's no loading image
Images do never need to be in a loading state.

When the user is sending images it can be displayed right away as he's
already having that image.
When the user is receiving an image we just don't display anything
until the image is ready to be displayed.

So, there's not requirement for a loading sequence.